+++ SerenityLib/Engine/src/ecs/ComponentLogic1.h
@@ -27,8 +27,8 @@ class ComponentLogic1 : public ComponentBaseClass {
     friend struct Engine::epriv::ComponentLogic1_SceneEnteredFunction;
     friend struct Engine::epriv::ComponentLogic1_SceneLeftFunction;
     private:
-        void*                               _userPtr;
-        boost::function<void(const float&)> _functor;
+        void*                                _userPtr;
+        boost::function<void(const double&)> _functor;
     public:
         ComponentLogic1(const Entity& _e) : ComponentBaseClass(_e) {
 			_userPtr = nullptr;
